fix(builder): embed the real Design panel in SiteEditorClient, matching Pixieset (TYN-315)
The previous fix just linked out to /design, but Pixieset embeds the
Template + accordion controls directly in the Website hub's sidebar, with
the live site preview already visible in the adjacent pane - no navigation
to a separate page.

- New app/builder/DesignPanelShared.tsx: extracts the accordion sections
  (Logo & Branding, Fonts, Colors, Animations, Spacing, Buttons) and the
  postMessage live-preview sync hook, shared between the standalone
  /design editor and this embedded panel so both stay identical
- SiteEditorClient's Design tab now renders those same controls inline;
  its existing persistent live-preview iframe gets ?__designPreview=1
  appended and receives real-time updates exactly like the standalone
  editor, unifying it into one view like Pixieset instead of two
- The shared Preview/Publish bar now branches per active tab: Design
  publishes the theme (/api/design/save), other tabs keep the original
  page-publish behavior
- app/builder/page.tsx fetches the initial theme server-side and passes
  it down, same pattern as the standalone /design page
